Changed the contract of Activity to provide Iterable<Renderable> for renderables rather than an array of Apologies, this commit is large and I really need to start leveraging git more by making smaller incremental commits.
Renderable. This makes it easier for the activity to construct its renderables without being bound by the rigid nature of arrays and their initial allocation size necessarily being known. The consumer of an activity only cares that it can iterate upon the renderables so as to draw them.

Improved the MainMenuActivity in subtle ways. Its still very much a work in progress.

Renamed the 'sendCommand' method of Activity to 'handle' instead, since that's clearer about the intent. Added method 'canHandle' so that a collection of Activity can accept command multicast broadcasts without needing the sender to know who is going to be able to handle it.

Renamed NumberUtil to NumberUtils since it's intended to be a pluralistic helper class.

Removed 'Point' from being used in RenderableText since it's an unnecessary abstraction (we turn x,y into a point and then turn a point back into x,y when we want to draw).

Work in progress on RenderableText to center it. This deserves more thinking.